+++ lib/src/judo_table.dart
@@ -35,7 +35,7 @@ class JudoTable extends StatelessWidget {
     this.alignment = Alignment.centerLeft,
   }) : super(key: key);
 
-  final int col;
+  final double col;
   final double row;
   final bool sortAscending;
   final int sortColumnIndex;
